[FIX] - Organise repository and update readme
