+++ README.md
@@ -1,14 +1,35 @@
-<img src="onetown.png" alt="One Town Over" width="400"/>
+# Our Texas Hold'em game (Team 1)
+
+![Alt text](readme_image.png)
+
 ## Team Members
 
-1. Nathan Donat Filliod <3
-2. Heber Trujillo Cruz
-3. Aaron Singh
-4. Jack Richards
+1. Heber Trujillo Cruz (Lead)
+2. Nathan Donat Filliod
+3. Jack Richards
 
-Compile server:
+## How to clone the repository
+```
+git clone git@github.com:sfsu-csc-667-spring-2025-roberts/term-project-one-town-over.git
+cd term-project-one-town-over
+```
 
-"npm run start" or "npm run start:dev"
+## How to compile backend
 
+```
+cd backend
+npm i
+npm run start:dev
+```
+
+## How to compile frontend
+
+```
+cd frontend
+```
+
+## Other links
+
+[Texas Hold'em DB Schema](https://dbdiagram.io/d/Texas-Holdem-Schema-67e9cd564f7afba184bdf8f6)
 
 [![Open in Codespaces](https://classroom.github.com/assets/launch-codespace-2972f46106e565e64193e422d61a12cf1da4916b45550586e14ef0a7c637dd04.svg)](https://classroom.github.com/open-in-codespaces?assignment_repo_id=18529720)
