+++ backend/src/server/views/root.ejs
@@ -1,0 +1,8 @@
+<html>
+    <head>
+        <title>Hello</title>
+    </head>
+    <body>
+    <p>Hello <%= name %></p>
+    </body>
+</html>
